refactor: move the fetch to contentful folder
